feat: export `fetchPackageManifest`
