+++ server/package.json
@@ -19,8 +19,6 @@
     "lru-cache": "^10.4.3",
     "nitropack": "^2.9.7",
     "npm-package-arg": "^11.0.2",
-    "ofetch": "^1.3.4",
-    "semver": "^7.6.2",
-    "ufo": "^1.5.3"
+    "semver": "^7.6.2"
   }
 }
