+++ shared/types.ts
@@ -42,3 +42,34 @@ export interface ResolvedPackageVersion extends Partial<PackageVersionMeta> {
   publishedAt: string | null
   lastSynced: number
 }
+
+export interface Packument {
+  'name': string
+  /**
+   * An object where each key is a version, and each value is the engines for
+   * that version.
+   */
+  'engines': Record<string, string>
+  /**
+   * An object where each key is a version, and each value is the manifest for
+   * that version.
+   */
+  'versions': Record<string, Omit<Packument, 'versions'>>
+  /**
+   * An object mapping dist-tags to version numbers. This is how `foo@latest`
+   * gets turned into `foo@1.2.3`.
+   */
+  'dist-tags': { latest: string } & Record<string, string>
+  /**
+   * Deprecated message for the package.
+   */
+  'deprecated'?: string
+  /**
+   * In the full packument, an object mapping version numbers to publication
+   * times, for the `opts.before` functionality.
+   */
+  'time': Record<string, string> & {
+    created: string
+    modified: string
+  }
+}
